+++ src/main/java/com/example/pet/repository/MemberRepository.java
@@ -3,8 +3,8 @@ package com.example.pet.repository;
 import com.example.pet.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
     public Member findByKakaoEmail(String kakaoEmail);
-    public Member findByMemberId(Long memberId);
+    public Member findByMemberId(int memberId);
     public Member findByNickname(String nickname);
 }
